requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
psutil>=5.9.0

# Web API (reemplaza Flask)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
PyJWT>=2.8.0
httpx>=0.27.0

# Networking / HTTP
requests>=2.31.0

# Audio / Voice
PyNaCl>=1.5.0
yt-dlp>=2024.11.6

# AI / Export
google-genai>=0.3.0
chat-exporter>=2.5.0

# Optional DB drivers (uncomment the one you use)
# psycopg2-binary>=2.9.0  # PostgreSQL
# PyMySQL>=1.1.0         # MariaDB / MySQL

# Optional / extras
# davey  # optional voice backend, uncomment if used
